Add a ScrollView
When the content of the taskLog becomes too big,
we need to scroll to be able to view the content
that's too far in the past.

ScrollView can have only one child, and although
we only had the TextEdit widget inside it, we
had to insert a LinearLayout to be able to properly
see and scroll the TextEdit.

The proper set of constraints was found by trial
and error.

Co-authored-by: Eric Herman <eric@freesa.org>